requirements.txt
tensorflow==2.10.0
tensorflow-addons
matplotlib
pillow
